backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
django-cors-headers==4.6.0
bcrypt==4.2.1
python-dotenv==1.0.1
gunicorn==23.0.0
